requirement.txt
click==7.1.2
configobj==4.7.2
contextlib2==0.6.0.post1
decorator==3.4.0
fail2ban==0.11.2
futures==3.1.1
importlib-resources==3.3.1
iniparse==0.4
iotop==0.6
joblib==0.14.1
mysql-connector-python==8.0.5
nltk==3.4.5
pathlib2==2.3.7.post1
perf==0.1
protobuf==3.17.3
pycurl==7.19.0
pygobject==3.22.0
pygpgme==0.3
pyliblzma==0.5.3
python-augeas==0.5.0
python-linux-procfs==0.4.9
pyudev==0.15
pyxattr==0.5.1
regex==2021.3.17
scandir==1.10.0
schedutils==0.4
singledispatch==3.7.0
six==1.9.0
slip==0.4.0
slip.dbus==0.4.0
tqdm==4.64.1
typing==3.10.0.0
urlgrabber==3.10
yum-metadata-parser==1.1.4
zipp==1.2.0
